feat: remove unwrap in favor of orThrow
removing `unwrap` in favor of `orThrow` to simplify the API design. This
simplifies the API design by removing `expect` and `unwrap` and
combining it into a singular `orThrow`, which takes a custom message as
an optional parameter. This simplifies the API by removing `expect` but
also makes it more clear to the develper as to what will happen if the
value is `None`. The trade-off is this no longer closely follows Rust's
Option. However, we think that trade off is valid as this would align
more with TypeScript developer's knowledge of exceptions.